pyproject.toml
[project]
name = "todo-app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Zyrian", email = "zyrian@mylocker.dev"},
]
dependencies = ["fastapi>=0.118.0", "tortoise-orm>=0.25.1", "uvicorn>=0.37.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
